pyproject.toml
[project]
name = "guest-the-song-community-ed"
version = "1.0.0"
authors = [
    { name = "Jose Cerrejon", email = "ulysess@gmail.com" }
]
description = "A music quiz game where players must guess songs from different categories"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "hid==1.0.7",
    "moviepy==2.1.2",
    "pygame==2.6.1",
    "pygame-menu==4.5.2",
    "pynput==1.8.1",
    "python-dotenv==1.1.0",
    "python-i18n==0.3.9",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "pytest==8.3.5",
    "ruff>=0.11.5",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
